src/test/java/seedu/address/model/person/MatricNumberTest.java
package seedu.address.model.person;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static seedu.address.testutil.Assert.assertThrows;

import org.junit.jupiter.api.Test;

public class MatricNumberTest {

    @Test
    public void constructor_null_throwsNullPointerException() {
        assertThrows(NullPointerException.class, () -> new MatricNumber(null));
    }

    @Test
    public void constructor_invalidMatricNumber_throwsIllegalArgumentException() {
        String invalidMatricNumber = "";
        assertThrows(IllegalArgumentException.class, () -> new MatricNumber(invalidMatricNumber));
    }

    @Test
    public void isValidMatricNumber() {
        // null matric number
        assertThrows(NullPointerException.class, () -> MatricNumber.isValidMatricNumber(null));

        // invalid matric numbers
        assertFalse(MatricNumber.isValidMatricNumber("")); // empty string
        assertFalse(MatricNumber.isValidMatricNumber(" ")); // spaces only
        assertFalse(MatricNumber.isValidMatricNumber("9123456")); // digits only
        assertFalse(MatricNumber.isValidMatricNumber("A123 4567Z")); // spaces within matric number
        assertFalse(MatricNumber.isValidMatricNumber("a1234567z")); // non-capital letters


        // valid phone numbers
        assertTrue(MatricNumber.isValidMatricNumber("A1234567Z")); // exactly 3 numbers
    }

    @Test
    public void equals() {
        MatricNumber matricNumber = new MatricNumber("A1234567Z");

        // same values -> returns true
        assertTrue(matricNumber.equals(new MatricNumber("A1234567Z")));

        // same object -> returns true
        assertTrue(matricNumber.equals(matricNumber));

        // null -> returns false
        assertFalse(matricNumber.equals(null));

        // different types -> returns false
        assertFalse(matricNumber.equals(5.0f));

        // different values -> returns false
        assertFalse(matricNumber.equals(new MatricNumber("A1234568Z")));
    }
}
